clearing_house: reduce User account size (#133)
* make order prices u64

* reduce base/quote to u64s in order

* make spot_balance u64

* make order_id a u32

* make 8 perp positions

* quote_asset_entry_amount as i64

* fix clearingHouse.ts test

* reduce base/quote asset amount on perp position

* shrink lp related data

* remove last_funding_rate_ts

* start fixing some of the sdk and tests

* tests passing

* reduce some casts

* tweak funding payments

* remove pnl from ts liqudiation records

* fix pre-commit hook

* all tests working

+++ programs/clearing_house/src/context.rs
@@ -445,18 +445,18 @@ pub struct OrderParams {
     pub market_type: MarketType,
     pub direction: PositionDirection,
     pub user_order_id: u8,
-    pub base_asset_amount: u128,
-    pub price: u128,
+    pub base_asset_amount: u64,
+    pub price: u64,
     pub market_index: u16,
     pub reduce_only: bool,
     pub post_only: bool,
     pub immediate_or_cancel: bool,
-    pub trigger_price: u128,
+    pub trigger_price: u64,
     pub trigger_condition: OrderTriggerCondition,
-    pub oracle_price_offset: i128,
+    pub oracle_price_offset: i64,
     pub auction_duration: Option<u8>,
     pub time_in_force: Option<u8>,
-    pub auction_start_price: Option<u128>,
+    pub auction_start_price: Option<u64>,
 }
 
 impl Default for OrderType {
